Releases/v1.6.0

+++ Lib/BrawlInstallerLib.py
@@ -1,4 +1,4 @@
-﻿version = "1.5.0"
+﻿version = "1.6.0"
 # BrawlInstallerLib
 # Functions used by BrawlInstaller plugins
 
@@ -701,7 +701,7 @@ def importStockIcons(cosmeticId, directory, tex0BresName, pat0BresName, rootName
 					# Color smash images in folders with multiple
 					if len(images) > 1 and not firstOnly:
 						writeLog("Color smashing stock icons")
-						ColorSmashImport(node, images, 256)
+						ColorSmashImport(node, images, 32)
 						writeLog("Imported color smashed icons")
 					elif len(images) >= 1:
 						writeLog("Importing standalone icon")
@@ -830,7 +830,7 @@ def addStockIcons(cosmeticId, images, position, tex0BresName, pat0BresName, root
 				# Import textures
 				if len(images) > 1:
 					writeLog("Color smashing stock icons")
-					ColorSmashImport(bresNode, images, 256)
+					ColorSmashImport(bresNode, images, 32)
 					writeLog("Imported color smashed icons")
 				else:
 					writeLog("Importing standalone icon")
@@ -3003,17 +3003,18 @@ def removeSong(songID, songDirectory='Victory!', tracklist='Results'):
 			# Get filename
 			path = MainForm.BuildPath + '/pf/sound/strm/' + songDirectory
 			directory = Directory.CreateDirectory(path)
-			brstmFile = getFileByName(childNode.SongFileName.split('/')[1] + ".brstm", directory)
-			# Back up song file
-			createBackup(brstmFile.FullName)
-			# Remove from tracklist
-			if childNode:
-				writeLog("Removing from" + tracklist + ".tlst")
-				childNode.Remove()
-			# Delete from directory
-			if brstmFile:
-				writeLog("Deleting file " + brstmFile.FullName)
-				brstmFile.Delete()
+			if '/' in childNode.SongFileName:
+				brstmFile = getFileByName(childNode.SongFileName.split('/')[1] + ".brstm", directory)
+				# Back up song file
+				createBackup(brstmFile.FullName)
+				# Remove from tracklist
+				if childNode:
+					writeLog("Removing from" + tracklist + ".tlst")
+					childNode.Remove()
+				# Delete from directory
+				if brstmFile:
+					writeLog("Deleting file " + brstmFile.FullName)
+					brstmFile.Delete()
 		BrawlAPI.SaveFile()
 		BrawlAPI.ForceCloseFile()
 		writeLog("Finished removing theme")
